## What changes were proposed in this pull request?

This pr contains:

1. in read path, accessing each element is heavy in Python, so modify it to reduce to access as possible.
2. in write path, every `list.append()` seems heavy, so capture the append method call in prior to use it.

In write path, we will be able to reduce to access each element more.

## How was this patch tested?

I did benchmark in my local environment.

before: 733 ms ± 43.1 ms
after: 413 ms ± 32.8 ms

but this is still slower than original (Pickle version).